Validações

+++ src/main/java/org/acme/services/ClienteService.java
@@ -1,5 +1,6 @@
 package org.acme.services;
 
+import io.quarkus.hibernate.orm.panache.PanacheEntityBase;
 import org.acme.Util.StringUtil;
 import org.acme.exceptions.ResponseBuilder;
 import org.acme.exceptions.ValidacaoException;
@@ -11,6 +12,7 @@ import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import javax.ws.rs.core.Response;
 import java.util.List;
+import java.util.Optional;
 
 @ApplicationScoped
 public class ClienteService extends Service {
@@ -24,9 +26,8 @@ public class ClienteService extends Service {
     public Response create(String json) {
         try {
             ClienteDTO clienteDTO = gson.fromJson(json, ClienteDTO.class);
-            validaCliente(clienteDTO);
+            validaCliente(clienteDTO,null,true);
             Cliente cliente = new Cliente();
-            cliente.setEndereco(new EnderecoNFE());
             fieldUtil.updateFieldsDtoToModel(cliente.getEndereco(), clienteDTO.getEndereco());
             clienteDTO.setEndereco(null);
             fieldUtil.updateFieldsDtoToModel(cliente, clienteDTO);
@@ -44,17 +45,34 @@ public class ClienteService extends Service {
 
 
 
-    public Cliente findOne(String uuid) {
-        return em.createQuery("SELECT c FROM Cliente c WHERE uuid = :uuid", Cliente.class)
-                .setParameter("uuid", uuid)
-                .getSingleResult();
+    public Response findOne(String uuid) {
+       try{
+           Cliente cliente = findOneEntity(uuid);
+           if (cliente != null) {
+               return Response.ok(cliente).build();
+           }
+           return Response.status(Response.Status.BAD_REQUEST).build();
+       }catch (Throwable t){
+           t.printStackTrace();
+           return ResponseBuilder.returnResponse();
+       }
+    }
+
+    private static Cliente findOneEntity(String uuid) {
+        try{
+            Optional<Cliente> cliente = Cliente.findByIdOptional(uuid);
+            return cliente.orElse(null);
+        }catch (Throwable t){
+            t.printStackTrace();
+            return null;
+        }
     }
 
     public Response update(String uuid, String json) {
         try {
-            Cliente cliente = findOne(uuid);
             ClienteDTO clienteDTO = gson.fromJson(json, ClienteDTO.class);
-
+            validaCliente(clienteDTO,null,true);
+            Cliente cliente = findOneEntity(uuid);
             em.merge(cliente);
             fieldUtil.updateFieldsDtoToModel(cliente, clienteDTO);
 
@@ -68,12 +86,24 @@ public class ClienteService extends Service {
     }
 
     public Response delete(String uuid) {
-        Cliente cliente = findOne(uuid);
-        em.remove(cliente);
-        return Response.ok().build();
+       try {
+           Cliente cliente =findOneEntity(uuid);
+           if(cliente != null){
+               em.remove(cliente);
+               return Response.ok(cliente).build();
+           }
+           throw new RuntimeException();
+       }catch (Throwable t){
+           t.printStackTrace();
+           return ResponseBuilder.returnResponse();
+       }
     }
-    private void validaCliente(ClienteDTO clienteDTO) {
-        ValidacaoException validacaoException = new ValidacaoException();
+    public void validaCliente(ClienteDTO clienteDTO,ValidacaoException validacaoException,boolean comEndereco) {
+        boolean validacaoExiste = validacaoException == null;
+        if(validacaoException == null){
+
+            validacaoException = new ValidacaoException();
+        }
 
         if(!StringUtil.stringValida(clienteDTO.getXNome())){
             validacaoException.add("Campo nome invalido");
@@ -81,8 +111,10 @@ public class ClienteService extends Service {
         if(!StringUtil.stringValida(clienteDTO.getCpfCnpj())){
             validacaoException.add("Campo CPF/CNPJ invalido");
         }
-        enderecoService.validaEndereco(validacaoException,clienteDTO.getEndereco(),true);
 
-        validacaoException.lancaErro();
+        enderecoService.validaEndereco(validacaoException,clienteDTO.getEndereco(),comEndereco);
+        if(validacaoExiste){
+            validacaoException.lancaErro();
+        }
     }
 }
